+++ src/js/header.js
@@ -1,17 +1,11 @@
 import "../styles/header.scss";
 import { createLogo } from "./stripes.js";
+import { loadNavbar } from'./nav.js';
 
-function createHeader() {
+export function createHeader() {
   createLogo();
+  loadNavbar();
 
-  const navButton = document.createElement("img");
-  navButton.classList.add("nav__button");
-  navButton.src = "/images/sliders/hamburger.png";
-  navButton.alt = "button for navbar";
-
-  document.querySelector(".header__container").prepend(navButton);
-
-  navButton.addEventListener("click", function () {});
 }
 
 createHeader();
